+++ src/features/filters/ui/Filters.stories.tsx
@@ -1,3 +1,5 @@
+import { OptionsType } from "@/configs/options";
+
 import Filters from "./Filters";
 
 import type { Meta, StoryObj } from "@storybook/react";
@@ -9,6 +11,13 @@ export default meta;
 
 type Story = StoryObj<typeof Filters>;
 
+const options: OptionsType = {
+    priceFrom: 0,
+    priceTo: 100000,
+    hits: false,
+    promotion: false,
+    inStock: false,
+};
 export const Default: Story = {
-    render: () => <Filters />,
+    render: () => <Filters options={options} setOptions={() => { }} />,
 };
