+++ src/shared/ui/productsPageLoader/styles.module.scss
