+++ src/shared/ui/subcategoryCard/SubcategoryCard.stories.tsx
@@ -11,6 +11,6 @@ type Story = StoryObj<typeof SubcategoryCard>;
 
 export const CardCategory: Story = {
     render: () => (
-        <SubcategoryCard image={""} text={"some text"} alt={" some text"} />
+        <SubcategoryCard title="Title" image={""} text={"some text"} alt={" some text"} />
     ),
 };
